requirements.txt
streamlit>=1.41.0
python-dotenv>=1.0.1
PyPDF2>=3.0.1
langchain>=0.3.14
langchain-community>=0.3.14
langchain-openai>=0.2.14
faiss-cpu>=1.9.0
tiktoken>=0.8.0
